remodified usertypes data type

+++ src/repositories/sql/userTypes/_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE usertypes (
-    "usertype_id" SERIAL NOT NULL PRIMARY KEY UNIQUE,
-    "permission_id" INT NOT NULL REFERENCES permissions(permission_id) DEFAULT 0,
+    "usertype_id" SMALLSERIAL NOT NULL PRIMARY KEY UNIQUE,
+    "permission_id" SMALLINT NOT NULL REFERENCES permissions(permission_id) DEFAULT 0,
     "usertype_title" VARCHAR(20) NOT NULL DEFAULT 'NO ACCESS',
     "created_at" TIMESTAMPTZ DEFAULT NOW(),
     "updated_at" TIMESTAMPTZ DEFAULT NOW()
